+++ src/rics/ml/__init__.py
@@ -1,0 +1,4 @@
+"""Stuff that's specific to working with machine learning.
+
+Additional dependencies may be required for some submodules.
+"""